+++ cmd/server/http_files/public/js/list.js
