+++ cmd/server/http_files/public/js/vue.js
