+++ cmd/server/settings.json
